Added required props: Buyer Category Discount Employee Order PaymnetTye Product Purchase Supplier
Adde new class:
ProductDetails

+++ OnlineShop/OnlineShop/Entities/ProductStocks.cs
@@ -3,7 +3,7 @@
     public class ProductStocks
     {
         public Product? product { get; set; }
-        public int ProductAmount { get; set; }
+        public decimal ProductAmount { get; set; }
         public ProductStocks()
         {
